+++ prettier.config.ts
@@ -1,0 +1,11 @@
+import type { Config } from 'prettier';
+
+const config: Config = {
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'all',
+  arrowParens: 'always',
+  printWidth: 100,
+};
+
+export default config;